test fix

+++ ledger/apply/application_test.go
@@ -539,8 +539,8 @@ func TestAppCallApplyCreate(t *testing.T) {
 	err = ApplicationCall(ac, h, &b, ad, &ep, txnCounter)
 	a.NoError(err)
 	br = b.putWithBalances[creator]
-	a.Equal(1, br.AppParams[appIdx].ExtraProgramPages)
-	a.Equal(1, br.TotalExtraAppPages)
+	a.Equal(int32(1), br.AppParams[appIdx].ExtraProgramPages)
+	a.Equal(int32(1), br.TotalExtraAppPages)
 }
 
 // TestAppCallApplyCreateOptIn checks balance record fields without tracking substages
@@ -1027,7 +1027,7 @@ func TestAppCallApplyDelete(t *testing.T) {
 	a.Equal(basics.AppParams{}, br.AppParams[appIdx])
 	a.Equal(basics.StateSchema{}, br.TotalAppSchema)
 	a.Equal(basics.EvalDelta{}, ad.EvalDelta)
-	a.Equal(0, br.TotalExtraAppPages)
+	a.Equal(int32(0), br.TotalExtraAppPages)
 }
 
 func TestAppCallApplyCreateClearState(t *testing.T) {
